+++ example/integration_test/plugin_integration_test.dart
@@ -10,14 +10,13 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
-import 'package:dnurse_health_plugin/dnurse_health_plugin.dart';
+import 'package:health_bridge/health_bridge.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('getPlatformVersion test', (WidgetTester tester) async {
-    final DnurseHealthPlugin plugin = DnurseHealthPlugin();
-    final String? version = await plugin.getPlatformVersion();
+    final String? version = await HealthBridge.getPlatformVersion();
     // The version string depends on the host platform running the test, so
     // just assert that some non-empty string is returned.
     expect(version?.isNotEmpty, true);