+++ lib/src/health_bridge_platform_interface.dart
@@ -1,22 +1,22 @@
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-import 'dnurse_health_plugin_method_channel.dart';
+import 'health_bridge_method_channel.dart';
 import 'models/health_data.dart';
 import 'models/health_platform.dart';
 
-/// DnurseHealthPlugin平台接口的抽象基类
-abstract class DnurseHealthPluginPlatform extends PlatformInterface {
+/// HealthBridge平台接口的抽象基类
+abstract class HealthBridgePlatform extends PlatformInterface {
   /// 构造函数
-  DnurseHealthPluginPlatform() : super(token: _token);
+  HealthBridgePlatform() : super(token: _token);
 
   static final Object _token = Object();
 
-  static DnurseHealthPluginPlatform _instance = MethodChannelDnurseHealthPlugin();
+  static HealthBridgePlatform _instance = MethodChannelHealthBridge();
 
   /// 当前平台实现的默认实例
-  static DnurseHealthPluginPlatform get instance => _instance;
+  static HealthBridgePlatform get instance => _instance;
 
   /// 设置平台实现实例，主要用于测试
-  static set instance(DnurseHealthPluginPlatform instance) {
+  static set instance(HealthBridgePlatform instance) {
     PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }